pyproject.toml
[project]
name = "harbortrainer"
version = "0.2.0"
description = "RL training for LLM agents on Harbor tasks using Prime-RL"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    # Prime-RL (includes verifiers)
    "prime-rl @ {root:uri}/prime-rl",
    
    # Core ML
    "torch",
    "numpy",
    "transformers",
    
    # Data handling
    "datasets",
    "blobfile",
    
    # Harbor (terminal/container execution)
    "docker",
    "harbor @ git+https://github.com/rishidesai/harbor.git",
    
    # Utilities
    "rich",
    "termcolor",
    "anyio",
    "pydantic>=2.0",
    "tomli",
    "openai",
    "litellm",
    
    # Logging
    "wandb",
    
    # Environment
    "python-dotenv",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "pyright",
]

[project.scripts]
# Entry point for harbor_env (for prime-rl to import)
harbor-train = "src.train:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
override-dependencies = [
    "openai>=1.106.1",
]
no-build-isolation-package = ["flash-attn"]

[tool.uv.sources]
torch = { index = "pytorch-cu128" }

[tool.uv.extra-build-dependencies]
flash-attn = [{ requirement = "torch", match-runtime = true }]

[tool.uv.extra-build-variables]
flash-attn = { FLASH_ATTENTION_SKIP_CUDA_BUILD = "TRUE" }

[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/test/cu128"
explicit = true

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
python_files = ["test_*.py", "*_test.py"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
ignore = ["E501"]

[tool.pyright]
pythonVersion = "3.12"
typeCheckingMode = "basic"
